requirements.txt
wheel==0.37.1
psycopg2==2.9.1
redis==4.0.2
pydantic[dotenv]==1.8.2
Flask==2.0.2
flask-restx==0.5.1
peewee==3.14.8
flask-jwt-extended==3.0.0
pyjwt==2.3.0
requests==2.26.0
jaeger-client==4.7.0
Flask-Opentracing==1.1.0
gunicorn==20.1.0
./auth_lib
